Migrate all calls to GURL::DeprecatedGetOriginAsURL()
This method will go away soon, so we get rid of Brave-specific
instances now to prepare for the future.

For now, this patch just makes sure we don't call the deprecated
method, but ideally we should probably make deeper changes not to
use a GURL as the type for origins, and use url::Origin instead.

Chromium change:

https://chromium.googlesource.com/chromium/src/+/e439902e17bbd1f540f7dc7f6e3e153871c76b34

commit e439902e17bbd1f540f7dc7f6e3e153871c76b34
Author: Matt Menke <mmenke@chromium.org>
Date:   Thu Oct 21 18:04:42 2021 +0000

    Reporting: Remove use of GURL::DeprecatedGetOriginAsURL().

    In some places this was called only to eventually convert the result
    to an Origin.

    Also make reporting APIs that were taking a URL only to convert it to
    an origin take origins instead.

    Bug: 512374