WIP: test Rewards.EnabledDuration P3A measurement.
Turn Brave Rewards on and off, checking for the appropriate
measurements. We hack the recorded timestamp to mock the passage
of time.

I've had trouble with the perference changes not propagating
within the timeframe of the test. This version overrides
`RewardsObserver::OnAdsEnabled` to keep track of the enabled
state, and adds a `WaitForAdsEnable` method spinning the message
loop, like InProcessBrowserTest::PostRunTestOnMainThread()
does at the end of the test, where tracing suggests the enable
callbacks otherwise finally run.

This is hacky, but seems to be enough to delay the test until the
enable state-change propagates to the ledger process and back to
OnPreferenceChanged and so to recording our measurement.

However, I haven't found a corresponding solution for propagating
disable.